Using DMA to sample multiple points at once

+++ src/Core/Inc/FastFouier.h
@@ -18,8 +18,8 @@ enum fourier_transform_direction { ftdFunctionToSpectrum, ftdSpectrumToFunction 
 //#define INCORRECT_SPECTRUM_SIZE_FOR_FFT				1
 //#define UNSUPPORTED_FTD								2
 
-void discreteFourierFast(const complex<double>* f, int i_max, complex<double>* F, fourier_transform_direction ftd);
-void discreteFourierFast2D(const complex<double>* f, int i_max, int j_max, complex<double>* F, fourier_transform_direction ftd);
+void discreteFourierFast(const complex<float>* f, int i_max, complex<float>* F, fourier_transform_direction ftd);
+void discreteFourierFast2D(const complex<float>* f, int i_max, int j_max, complex<float>* F, fourier_transform_direction ftd);
